+++ NEWS.md
@@ -51,6 +51,10 @@
 
 * Add support for `dplyr::tally()` (#203).
 
+* `count()` and `add_count()` now warn or error when argument `wt` is used
+  since it is not supported. The behavior depends on the global option
+  `tidypolars_unknown_args` (#204).
+
 ## Bug fixes
 
 * Fix behavior of `mutate()` and `summarize()` when they don't contain any